project/DB/serializers.py
from rest_framework import serializers
from DB.models import User, Dictionary

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = '__all__'

class DictionarySerializer(serializers.ModelSerializer):
    class Meta:
        model = Dictionary
        fields = '__all__' 

